+++ eras/dijkstra/impl/src/Cardano/Ledger/Dijkstra/State/Account.hs
@@ -3,7 +3,6 @@
 
 module Cardano.Ledger.Dijkstra.State.Account () where
 
-import Cardano.Ledger.BaseTypes
 import Cardano.Ledger.Conway.State
 import Cardano.Ledger.Dijkstra.Era
 import qualified Data.Map.Strict as Map
@@ -17,17 +16,17 @@ instance EraAccounts DijkstraEra where
 
   accountsMapL = lens caStates $ \cas asMap -> cas {caStates = asMap}
 
-  balanceAccountStateL = lens casBalance $ \cas b -> cas {casBalance = b}
+  balanceAccountStateL = balanceConwayAccountStateL
+  {-# INLINE balanceAccountStateL #-}
 
-  depositAccountStateL = lens casDeposit $ \cas d -> cas {casDeposit = d}
+  depositAccountStateL = depositConwayAccountStateL
+  {-# INLINE depositAccountStateL #-}
 
-  stakePoolDelegationAccountStateL =
-    lens (strictMaybeToMaybe . casStakePoolDelegation) $ \cas d ->
-      cas {casStakePoolDelegation = maybeToStrictMaybe d}
+  stakePoolDelegationAccountStateL = stakePoolDelegationConwayAccountStateL
+  {-# INLINE stakePoolDelegationAccountStateL #-}
 
   unregisterAccount = unregisterConwayAccount
 
 instance ConwayEraAccounts DijkstraEra where
-  dRepDelegationAccountStateL =
-    lens (strictMaybeToMaybe . casDRepDelegation) $ \cas d ->
-      cas {casDRepDelegation = maybeToStrictMaybe d}
+  dRepDelegationAccountStateL = dRepDelegationConwayAccountStateL
+  {-# INLINE dRepDelegationAccountStateL #-}