+++ libs/cardano-ledger-binary/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Version history for `cardano-ledger-binary`
 
+## 1.8.1.0
+
+* Add `DecShareCBOR` instance for `VMap VB VS`
+
 ## 1.8.0.0
 
 * Add `Uniform`, `UniformRange` instances and fix `Random` instance