+++ libs/cardano-ledger-core/CHANGELOG.md
@@ -1,5 +1,10 @@
 # Version history for `cardano-ledger-core`
 
+## 1.20.0.0
+
+* Add `Storable` instance for `NonZero`, `CompactForm Coin`, `KeyHash`, `ScriptHash`, `Credential` and `StakeWithDelegation`
+* Switch `ActiveStake` to use `VS` for the value in the `VMap`
+
 ## 1.19.0.0
 
 * In `SnapShot`, subsume `ssDelegations` into `ssActiveStake`.